- Simbols added

+++ WeatherApp/Data.cs
@@ -32,18 +32,19 @@ namespace WeatherApp
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
+        public double pressure { get; set; }
         public int humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
         public String foto { get; set; }
 
+   
     }
 
     public class Wind
     {
         public double speed { get; set; }
-        public int deg { get; set; }
+        public double deg { get; set; }
     }
 
     public class Clouds
